% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{get_group_dots_data}
\alias{get_group_dots_data}
\title{Finds where the dots for a specific group are located}
\usage{
get_group_dots_data(dots_data, i, groups)
}
\arguments{
\item{dots_data}{input data}

\item{i}{index of the current group}

\item{groups}{labels of all groups in the plot}
}
\value{
dataframe with positions
}
\description{
Returns a dataframe with positions of samples that can be superimposed on the area plot.
}
