% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{loreplotr}
\alias{loreplotr}
\title{Draw Loreplot}
\usage{
loreplotr(df, x, y, draw_dots = TRUE)
}
\arguments{
\item{df}{The input data}

\item{x}{Continuous variable to be shown on the x-axis}

\item{y}{Categorical variable, predicated probabilities shown on y-axis}
}
\value{
ggplot2 plot
}
\description{
Draws a logistic regression plot for the provided data
}
\examples{
mtcars \%>\% loreplotr("mpg", "cyl")
}
