% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{plot_area}
\alias{plot_area}
\title{Plot Area}
\usage{
plot_area(
  df,
  x,
  y,
  draw_dots = TRUE,
  dots_shape = 21,
  dots_fill = "white",
  dots_colour = "white",
  dots_size = 3,
  dots_alpha = 0.7
)
}
\arguments{
\item{df}{The input data}

\item{x}{Continuous variable to be shown on the x-axis}

\item{y}{Categorical variable, predicated probabilities shown on y-axis (if there are only two classes they need to be 0 and 1 !)}

\item{draw_dots}{Show a dot in the plot for eachs sample (default=TRUE)}

\item{dots_shape}{Shape of the dots (default=21)}

\item{dots_fill}{Fill color of the dots (default="white")}

\item{dots_colour}{Border color of the dots (default="white")}

\item{dots_size}{Size of the dots (default=3)}

\item{dots_alpha}{Transparency of the dots (default=0.7)}
}
\value{
ggplot2 plot
}
\description{
Internal loreplotr function that draws the plot
}
