% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{loreplotr}
\alias{loreplotr}
\title{Draw Loreplot}
\usage{
loreplotr(
  df,
  x,
  y,
  draw_dots = TRUE,
  dots_shape = 21,
  dots_fill = "white",
  dots_colour = "white",
  dots_size = 3,
  dots_alpha = 0.7
)
}
\arguments{
\item{df}{The input data}

\item{x}{Continuous variable to be shown on the x-axis}

\item{y}{Categorical variable, predicated probabilities shown on y-axis}

\item{draw_dots}{Show a dot in the plot for each sample (default=TRUE)}

\item{dots_shape}{Shape of the dots (default=21)}

\item{dots_fill}{Fill color of the dots (default="white")}

\item{dots_colour}{Border color of the dots (default="white")}

\item{dots_size}{Size of the dots (default=3)}

\item{dots_alpha}{Transparency of the dots (default=0.7)}
}
\value{
ggplot2 plot
}
\description{
Draws a logistic regression plot for the provided data
}
\examples{
library(dplyr)
library(loreplotr)
data("mtcars")

mtcars$cyl <- paste("cyl", mtcars$cyl, sep="_")
g <- loreplotr(mtcars, "mpg", "cyl")
g
}
