% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{loreplotr}
\alias{loreplotr}
\title{Draw Loreplot}
\usage{
loreplotr(df, x, y, draw_dots = TRUE)
}
\arguments{
\item{df}{The input data}

\item{x}{Continuous variable to be shown on the x-axis}

\item{y}{Categorical variable, predicated probabilities shown on y-axis}

\item{draw_dots}{Show a dot in the plot for each sample (default=TRUE)}
}
\value{
ggplot2 plot
}
\description{
Draws a logistic regression plot for the provided data
}
\examples{
library(dplyr)
library(loreplotr)
data("mtcars")

mtcars$cyl <- paste("cyl", mtcars$cyl, sep="_")
g <- loreplotr(mtcars, "mpg", "cyl")
g
}
