% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loreplotr.R
\name{plot_area}
\alias{plot_area}
\title{Plot Area}
\usage{
plot_area(df, x, y, draw_dots = TRUE)
}
\arguments{
\item{df}{The input data}

\item{x}{Continuous variable to be shown on the x-axis}

\item{y}{Categorical variable, predicated probabilities shown on y-axis}

\item{draw_dots}{Show a dot in the plot for eachs sample (default=TRUE)}
}
\value{
ggplot2 plot
}
\description{
Internal loreplotr function that draws the plot
}
